requirements.txt
polars
python-dotenv
pydantic
pydantic-settings
pymongo
requests
motor
python-multipart
numpy
pillow
tensorflow==2.16.1
torch
torchvision
ultralytics
keras==3.3.3
transformers
opencv-python
minio
ijson